Update style.css
Made these changes in this commit:
Give the link, visited, focus, and hover states some colors that go with the color of the horizontal bars at the top and bottom of the page.
Make it so that links are underlined by default, but when you hover or focus them the underline disappears.